实现打卡！

+++ src/main/java/com/software/personalmanagement/entity/Attendance.java
@@ -2,8 +2,6 @@ package com.software.personalmanagement.entity;
 
 import lombok.Data;
 
-//打卡类
-
 @Data
 public class Attendance {
     private Integer id;
@@ -11,6 +9,6 @@ public class Attendance {
     private String name;
     private String startTime;       //上班打卡
     private String endTime;         //下班打卡
-    private Double workTime;        //工作时长
+    private Integer workTime;        //工作时长
     private Integer status;
 }
