+++ src/main/java/com/software/personalmanagement/vo/AttendanceVO.java
@@ -2,8 +2,6 @@ package com.software.personalmanagement.vo;
 
 import lombok.Data;
 
-import java.util.Date;
-
 @Data
 public class AttendanceVO {
     private Integer id;
@@ -11,6 +9,6 @@ public class AttendanceVO {
     private String name;
     private String startTime;       //上班打卡
     private String endTime;         //下班打卡
-    private Double workTime;        //工作时长
+    private Integer workTime;        //工作时长
     private String status;
 }
